+++ DTCD-StyleSystem/src/base-components/BaseSelect/BaseSelect.js
@@ -13,6 +13,7 @@ export default class BaseSelect extends HTMLElement {
   #value;
   #itemSlot;
   #opened = false;
+  #doValidation = false;
 
   static get observedAttributes() {
     return [
@@ -74,7 +75,7 @@ export default class BaseSelect extends HTMLElement {
     this.#header.innerHTML = newValue;
     this.#searchInput.setAttribute('placeholder', newValue);
 
-    this.validate();
+    this.#doValidation && this.validate();
 
     this.dispatchEvent(new Event('input'));
     if (oldValue !== this.value) {
@@ -159,7 +160,7 @@ export default class BaseSelect extends HTMLElement {
     }
 
     this.#message.innerHTML = this.#invalid && this.#messageText ? this.#messageText : '';
-    this.#message.style.display = this.#message.textContent.length ? '' : 'none';
+    this.#message.style.padding = this.#message.textContent.length ? '' : '0';
   }
 
   validate() {
@@ -222,13 +223,15 @@ export default class BaseSelect extends HTMLElement {
         this.#searchInput.value = '';
       }
 
-      this.validate();
+      this.#doValidation && this.validate();
     }
 
     return this.#opened;
   }
 
   connectedCallback() {
+    this.#doValidation = true;
+
     this.#fieldWrapper.addEventListener('click', this.#handleFieldWrapperClick);
     this.#searchInput.addEventListener('input', this.#handleSearchFieldInput);
   }
@@ -241,7 +244,7 @@ export default class BaseSelect extends HTMLElement {
   attributeChangedCallback(attrName, oldValue, newValue) {
     switch (attrName) {
       case 'disabled':
-        if (newValue) {
+        if (this.disabled) {
           this.#searchInput.setAttribute('disabled', true);
           this.#selectContainer.classList.add('disabled');
         } else {