+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0]
+
+### Added
+
+- enabling check BaseInput, BaseTextarea, BaseSelect after connection to page
+- label in BaseSwitch
+
+### Changed
+
+- 'display:none' with 'padding:0' in error message of BaseInput, BaseTextarea and BaseSelect
+- setting boolean attributes (required, etc.)
+
+### Fixed
+
+- synced typed value in datetime picker
+- long value in BaseSelect component
+- DataSourceSelect create button not work
+
 ## [0.6.0]
 
 ### Added